checks/log4J/requirements.txt
requests
PyCryptodome
dnslib
